requirements.txt
pyyaml
setuptools
scipy
numpy